resolve comments

+++ mockgcp/mockcompute/instancetemplate_v1.go
@@ -12,32 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Copyright 2024 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 package mockcompute
 
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/emptypb"
-	"google.golang.org/protobuf/types/known/timestamppb"
 
 	pb "github.com/GoogleCloudPlatform/k8s-config-connector/mockgcp/generated/mockgcp/cloud/compute/v1"
 )
@@ -54,34 +38,34 @@ func (s *InstanceTemplatesV1) Insert(ctx context.Context, req *pb.InsertInstance
 	}
 
 	fqn := "projects/" + req.GetProject() + "/global/instanceTemplates/" + name
-	selfLink := fmt.Sprintf("https://www.googleapis.com/compute/v1/projects/%s/global/instanceTemplates/%s", req.GetProject(), name)
+	selfLink := fmt.Sprintf("https://compute.googleapis.com/compute/v1/projects/%s/global/instanceTemplates/%s", req.GetProject(), name)
 
 	obj := proto.Clone(req.GetInstanceTemplateResource()).(*pb.InstanceTemplate)
 	obj.SelfLink = PtrTo(selfLink)
 	obj.Kind = PtrTo("compute#instanceTemplate")
-	obj.Id = PtrTo(uint64(time.Now().UnixNano()))
-	obj.CreationTimestamp = PtrTo(timestamppb.Now().String())
+	obj.Id = PtrTo(s.generateID())
+	obj.CreationTimestamp = PtrTo(s.nowString())
 	obj.Name = PtrTo(name)
 
 	if obj.Properties == nil {
 		obj.Properties = &pb.InstanceProperties{}
 	}
-	if len(obj.Properties.Disks) > 0 {
-		if obj.Properties.Disks[0].DeviceName == nil {
-			obj.Properties.Disks[0].DeviceName = PtrTo("boot")
+	for _, disk := range obj.Properties.Disks {
+		if disk.DeviceName == nil {
+			disk.DeviceName = PtrTo("boot")
 		}
-		if obj.Properties.Disks[0].Index == nil {
-			obj.Properties.Disks[0].Index = PtrTo(int32(0))
+		if disk.Index == nil {
+			disk.Index = PtrTo(int32(0))
 		}
-		obj.Properties.Disks[0].Kind = PtrTo("compute#attachedDisk")
+		disk.Kind = PtrTo("compute#attachedDisk")
 	}
-	if len(obj.Properties.NetworkInterfaces) > 0 {
-		if obj.Properties.NetworkInterfaces[0].Name == nil {
-			obj.Properties.NetworkInterfaces[0].Name = PtrTo("nic0")
+	for _, networkInterface := range obj.Properties.NetworkInterfaces {
+		if networkInterface.Name == nil {
+			networkInterface.Name = PtrTo("nic0")
 		}
-		obj.Properties.NetworkInterfaces[0].Kind = PtrTo("compute#networkInterface")
-		if len(obj.Properties.NetworkInterfaces[0].AccessConfigs) > 0 {
-			obj.Properties.NetworkInterfaces[0].AccessConfigs[0].Kind = PtrTo("compute#accessConfig")
+		networkInterface.Kind = PtrTo("compute#networkInterface")
+		for _, accessConfig := range networkInterface.AccessConfigs {
+			accessConfig.Kind = PtrTo("compute#accessConfig")
 		}
 	}
 	if obj.Properties.Scheduling == nil {
@@ -100,7 +84,7 @@ func (s *InstanceTemplatesV1) Insert(ctx context.Context, req *pb.InsertInstance
 		obj.Properties.Metadata = &pb.Metadata{}
 	}
 	if obj.Properties.Metadata.Fingerprint == nil {
-		obj.Properties.Metadata.Fingerprint = PtrTo("nEPyOfWs0II=")
+		obj.Properties.Metadata.Fingerprint = PtrTo(computeFingerprint(obj.Properties.Metadata))
 	}
 	if obj.Properties.Metadata.Kind == nil {
 		obj.Properties.Metadata.Kind = PtrTo("compute#metadata")
@@ -122,6 +106,9 @@ func (s *InstanceTemplatesV1) Insert(ctx context.Context, req *pb.InsertInstance
 }
 
 func (s *InstanceTemplatesV1) Get(ctx context.Context, req *pb.GetInstanceTemplateRequest) (*pb.InstanceTemplate, error) {
+	if req.GetInstanceTemplate() == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "InstanceTemplate is required")
+	}
 	fqn := "projects/" + req.GetProject() + "/global/instanceTemplates/" + req.GetInstanceTemplate()
 	obj := &pb.InstanceTemplate{}
 	if err := s.storage.Get(ctx, fqn, obj); err != nil {
@@ -131,6 +118,9 @@ func (s *InstanceTemplatesV1) Get(ctx context.Context, req *pb.GetInstanceTempla
 }
 
 func (s *InstanceTemplatesV1) Delete(ctx context.Context, req *pb.DeleteInstanceTemplateRequest) (*pb.Operation, error) {
+	if req.GetInstanceTemplate() == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "InstanceTemplate is required")
+	}
 	fqn := "projects/" + req.GetProject() + "/global/instanceTemplates/" + req.GetInstanceTemplate()
 	deleted := &pb.InstanceTemplate{}
 	if err := s.storage.Delete(ctx, fqn, deleted); err != nil {
@@ -138,7 +128,8 @@ func (s *InstanceTemplatesV1) Delete(ctx context.Context, req *pb.DeleteInstance
 	}
 
 	op := &pb.Operation{
-		OperationType: PtrTo("compute.instanceTemplates.delete"),
+		// Different than the other CRUD operation, DELETE operation type does not have "compute.instanceTemplates."
+		OperationType: PtrTo("delete"),
 		TargetLink:    deleted.SelfLink,
 		TargetId:      deleted.Id,
 		User:          PtrTo("user@example.com"),
